Add BackgroundType with options of Transparent and Solid (#22)
Adds a new field to Style to select, Background Type: lets athletes choose to have their art with a transparent background or a solid color background. In the future, this will be used for multiple different gradient backgrounds too.
Adds feature education next to transparent choice and in the preview menu when a checkered background is shown.
Update the about screen to provide a link to personal e-mail.